frontend/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
pdfplumber
python-dotenv